Initial virtual screen "Record Studio"
This is a workaround for the following issue:

https://github.com/xmonad/xmonad/issues/412

Where it would not be possible to capture window on another,
unfocused workspace with programs like OBS or gstreamer.

With the key combination Mod+Backspace one can mark a workspace for
recording, to set it "On the Air". This ensures that when the marked
workspace is not focused that it can still be captured.

The hack to ensure this: An unfocused workspace is moved to an unused
screen, here HDMI-1.

Current known issues:

* A few frames might leak from an unmarked workspace when switching
  from/to it.
* HDMI-1 won't be available for normal usage. HDMI-1 and the resolution
  are currently hardcoded.
* With OBS the "virtual" screen HDMI-1 needs a fixed position, here 0x0.
  So it needs to be left of the real screen to ensure that.
  Because OBS when using a screen capture will not update the position
  automatically and if the real screen and virtual screen were swapping
  back and forth for position 0x0, this could lead to OBS capturing a
  screen with a workspace you don't actually want to share.
... and many more, see TODOs in the code.