terminal: switch from lxterminal to xfce4-terminal
Neither gnome-terminal nor lxterminal seem to support a
"--disable-factory" option anymore, creating a new window will
typically not use a new process.

This breaks the usage of "xprop _NET_WM_PID" and "spawnOn" to spawn a
terminal on a specific workspace.

xfce4-terminal however still has such an option and "spawnOn" still
works with it. Therefore switching to xfce4-terminal.

Spawning a new terminal is now a bit slower, but hopefully we'll get
used to it...